Fixed Advection Stage, 2D Fluid Simulation Works!
+ Turns out I wasn't using the advection coordinate I was calculating.
+ Still not massivley stable, there seems to be certain magic numbers to get the simulation to behave correctly.

+++ Assets/Detonate-FSVE/2D/Classes/FluidSim2DParams.cs
@@ -8,7 +8,7 @@ public class FluidSim2DParams
     [Space] [Header("Simulation")]
     public int width = 512;
     public int height = 512;
-    public float jacobi_iterations = 40;
+    public uint jacobi_iterations = 40;
 
     [Header("Smoke")]
     public float smoke_buoyancy = 1.0f;
